+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 * pool: cleanup relay `Error` variants ([Yuki Kishimoto])
 * pool: acquire service watcher receiver outside the auto-connect loop ([Yuki Kishimoto])
 * pool: decrease `MAX_RETRY_INTERVAL` to 60 secs ([Yuki Kishimoto])
+* pool: rework retry interval calculation ([Yuki Kishimoto])
 * sdk: refactor POW difficulty management ([Yuki Kishimoto])
 
 ### Added